Fix compilation warnings on device

+++ apps/code/clipboard.h
@@ -12,7 +12,7 @@ public:
 
   static Clipboard * sharedClipboard();
   static const UTF8Helper::TextPair * PythonTextPairs();
-  static bool ShouldReplaceLetterE(const char * text, int length, int position);
+  static bool ShouldReplaceLetterE(const char * text, size_t length, size_t position);
   void enterPython() { replaceCharForPython(true); }
   void exitPython() { replaceCharForPython(false); }
 private:
